Edited the profile of EnchantedBookMaker.make(), short are now converted inside the method

+++ src/main/java/nemexlib/api/items/vanilla/EnchantedBookMaker.java
@@ -7,12 +7,12 @@ import net.minecraft.nbt.NBTTagList;
 
 public class EnchantedBookMaker {
 
-    public static ItemStack make(short level, short id) {
+    public static ItemStack make(int level, int id) {
         NBTTagCompound nbt = new NBTTagCompound(), tags = new NBTTagCompound();
         NBTTagList nbtList = new NBTTagList();
         nbt.setTag("StoredEnchantments", nbtList);
-        tags.setShort("lvl", level);
-        tags.setShort("id", id);
+        tags.setShort("lvl", (short) level);
+        tags.setShort("id", (short) id);
         nbtList.appendTag(tags);
         ItemStack book = new ItemStack(Items.enchanted_book);
         book.setTagCompound(nbt);
